+++ CHANGES.md
@@ -1,12 +1,23 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [UNRELEASED]
+
 ### Changed
+
+- Identifier tokens now never have a `ContextualKind` by @GGG-KILLER in https://github.com/LorettaDevs/Loretta/pull/134.
 - **[Breaking]** `LuaSyntaxOptions.Luau` has been modified to enable floor division by @GGG-KILLER in https://github.com/LorettaDevs/Loretta/pull/133.
+
+### Fixed
+
+- Fixed `continue` still being parsed as a keyword even when `ContinueType` was `ContinueType.None` by @GGG-KILLER
+  in https://github.com/LorettaDevs/Loretta/pull/134;
+- Fixed `goto` and `::label::` still being parsed when `AcceptGoto` was `false` by @GGG-KILLER in
+  https://github.com/LorettaDevs/Loretta/pull/134.
 
 ## v0.2.12
 ### Fixed